Work on simulation and making grammar generation simpler
