+++ runtime/Rust/Cargo.toml
@@ -19,11 +19,3 @@ debug = []
 
 [dependencies]
 lazy_static = "^1.4"
-uuid = "1.16.0"
-byteorder = "^1"
-murmur3 = "=0.4" # 0.5 is incompatible currently
-bit-set = "0.8.0"
-once_cell = "^1.2"
-typed-arena = "^2.0"
-better_any = "0.2.0"
-parking_lot = "0.12"
