+++ runtime/Rust/src/atn/rule.rs
@@ -3,7 +3,7 @@ use std::fmt::Debug;
 use std::fmt::Formatter;
 use std::fmt::Result;
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub enum ATNRule {
     Lexer {
         start_state: ATNStateRef,
