+++ runtime/Rust/src/lex/action.rs
@@ -25,7 +25,7 @@ impl LexerActionType {
     }
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub enum LexerAction {
     Channel(usize),
     Custom {
