+++ runtime/Rust/src/lex/mod.rs
@@ -2,5 +2,6 @@ mod action;
 pub mod token;
 mod lexer;
 
+pub use lexer::Lex;
 pub use lexer::Lexer;
 pub use action::LexerAction;
